fix the activity stream code
Previously, it would check for a list of types and IDs, but not match them up, so multiple objects with the same type or same ID would get included in the list. Now, it ends up doing something like “(id = first_id AND type = first_type) OR (id = second_id AND type = second_type) OR…” for every object in the shared list.