New tag managment - new tags can be introduced which need moderation.
New tags are saved as ManagedTags which can be approved by staff.
Approving a ManagedTag creates an equivalent Taggit.Tag and applies
it everywhere the ManagedTag was applied.  The approved ManagedTag
is then removed from use.
If the new_tag form field is used to enter a tag which has already
been approved, it will be automatically promoted to the normal
TaggitTag in the tags field.
This has been tested on CaseReports, Bibliography References,
Discussions and Documents.
Additionally, the migrations predefine a number of tags for use.
refs #185 #876