Invite co-author to join site refs #855
Send co-author an email with a custom link to register with their
existing user. Set the registration form to take their user id, and
autofill their email address. Don't allow the user to change the email
at this point, because that will create a separate member not associated
with the co-author account. Update the registration form to allow
existing users to register, as long as their account is not active.
Update the account name based on field entered in the registration form.