inverse meaning of opt_out_of_email refs #1089 #1047
Update the opt_out_of_email field label and negate the condition
everywhere it is checked. Yes, this is confusing on the backend, but it
is temporary as a new field will be replacing this soon (already fixed in
commit 55c8bb7). This means that all users now have emails disabled without
need of a migration.

This should not be merged to master, but only used until the other
changes are ready for prod.