CaseReports to store a 'frozen' copy of their metadata.    refs #1288
Supplants action.send with a wrapper that detects storable metadata
and appends that to the send call as a 'frozen' copy of the object.

Object need merely has a to_dict method.

Also wrapped some exception handling around some of the outgoing email
sends that were failing in my local environment so that CRs can transition
between states even if we cant send email.