Invites by moderators bypasss closed group joining workflow.  refs #1432
Additions to closed groups by moderators are recorded in the approver field of the membership.
Project.add_member can now also pass an initial membership state.

For non-members, the account is precreated along with membership.
When they complete their registration, their memberships will be already approved
and waiting for them.