Suppress activity feed enties that are missing the action object.  refs #1438
See also #1461 and #1441.     I'm thinking these were invites
to people that didn't yet exist as users or something like that.
The verb on the actions was 'joined' but there was no object
indicating who.   The actors were all Sarah.

Since we are not displaying these entries ( in order to avoid the
issue with repeated horizontal rules being displayed with nothing
in between ), I've added an html comment which should point us
directly to the actions by their id.